+++ src/linear_proof.rs
@@ -73,20 +73,20 @@ impl LinearProof {
         assert!(n.is_power_of_two());
 
         transcript.innerproduct_domain_sep(n as u64);
-        transcript.append_point(b"C", &C);
+        transcript.append_point(b"C", C);
 
         let lg_n = n.next_power_of_two().trailing_zeros() as usize;
         let mut L_vec = Vec::with_capacity(lg_n);
         let mut R_vec = Vec::with_capacity(lg_n);
 
         while n != 1 {
-            n = n / 2;
+            n /= 2;
             let (a_L, a_R) = a.split_at_mut(n);
             let (b_L, b_R) = b.split_at_mut(n);
             let (G_L, G_R) = G.split_at_mut(n);
 
-            let c_L = inner_product(&a_L, &b_R);
-            let c_R = inner_product(&a_R, &b_L);
+            let c_L = inner_product(a_L, b_R);
+            let c_R = inner_product(a_R, b_L);
 
             let s_j = Scalar::random(rng);
             let t_j = Scalar::random(rng);
@@ -116,9 +116,9 @@ impl LinearProof {
 
             for i in 0..n {
                 // a_L = a_L + x_j^{-1} * a_R
-                a_L[i] = a_L[i] + x_j_inv * a_R[i];
+                a_L[i] += x_j_inv * a_R[i];
                 // b_L = b_L + x_j * b_R
-                b_L[i] = b_L[i] + x_j * b_R[i];
+                b_L[i] += x_j * b_R[i];
                 // G_L = G_L + x_j * G_R
                 G_L[i] = RistrettoPoint::vartime_multiscalar_mul(
                     &[Scalar::one(), x_j],
@@ -149,6 +149,7 @@ impl LinearProof {
         }
     }
 
+    /// Verify a linear proof
     pub fn verify(
         &self,
         n: usize,
@@ -165,7 +166,7 @@ impl LinearProof {
         b_vec: Vec<Scalar>,
     ) -> Result<(), ProofError> {
         transcript.innerproduct_domain_sep(n as u64);
-        transcript.append_point(b"C", &C);
+        transcript.append_point(b"C", C);
         let (x_vec, x_inv_vec, b_0) = self.verification_scalars(n, transcript, b_vec)?;
 
         transcript.append_point(b"S", &self.S);
@@ -249,10 +250,10 @@ impl LinearProof {
             transcript.validate_and_append_point(b"R", R)?;
             let x_j = transcript.challenge_scalar(b"x_j");
             challenges.push(x_j);
-            n_mut = n_mut / 2;
+            n_mut /= 2;
             let (b_L, b_R) = b.split_at_mut(n_mut);
             for i in 0..n_mut {
-                b_L[i] = b_L[i] + x_j * b_R[i];
+                b_L[i] += x_j * b_R[i];
             }
             b = b_L;
         }