+++ src/util.rs
@@ -102,8 +102,8 @@ impl VecPoly1 {
     pub fn eval(&self, x: Scalar) -> Vec<Scalar> {
         let n = self.0.len();
         let mut out = vec![Scalar::zero(); n];
-        for i in 0..n {
-            out[i] = self.0[i] + self.1[i] * x;
+        for (i, item) in out.iter_mut().enumerate().take(n) {
+            *item = self.0[i] + self.1[i] * x;
         }
         out
     }
@@ -225,9 +225,9 @@ pub fn scalar_exp_vartime(x: &Scalar, mut n: u64) -> Scalar {
     while n > 0 {
         let bit = n & 1;
         if bit == 1 {
-            result = result * aux;
+            result *= aux;
         }
-        n = n >> 1;
+        n >>= 1;
         aux = aux * aux; // FIXME: one unnecessary mult at the last step here!
     }
     result
@@ -250,7 +250,7 @@ pub fn sum_of_powers(x: &Scalar, n: usize) -> Scalar {
     while m > 2 {
         factor = factor * factor;
         result = result + factor * result;
-        m = m / 2;
+        m /= 2;
     }
     result
 }